+++ test/fixtures/primary_constructors/const_class.input.dart
@@ -1,0 +1,6 @@
+class Point {
+  final int x;
+  final int y;
+
+  const Point(this.x, this.y);
+}